menambahkan aksi delete pada bagian feedback dan laporan di halaman s…

+++ app/Views/admin/news.php
@@ -44,7 +44,7 @@
                             <td>
                                 <button type="button" class="btn btn-secondary btn-sm">Edit</button>
                                 <!-- <button type="button" class="btn btn-primary btn-sm">Hapus</button> -->
-                                <a href="/news/delete/<?= $r['id_news']; ?>" class="btn btn-primary btn-sm">Delete</a>
+                                <a href="/news/delete/<?= $r['id_news']; ?>" class="btn btn-primary btn-sm" onclick="return confirm('Yakin data ini akan dihapus');">Delete</a>
                                 <a href="/news/detailNews/<?= $r['id_news']; ?>" class="btn btn-info btn-sm">Detail</a>
                             </td>
                         </tr>
